now we have dusk & dawn and variing view distance!

+++ dangerdeep/src/global_data.h
@@ -24,7 +24,9 @@ void deinit_global_data(void);
 
 void get_date(double t, unsigned& year, unsigned& month, unsigned& day);
 double get_time(unsigned year, unsigned month, unsigned day);
-unsigned get_day_time(double t);
+// computes part of day (0 night, 1 sunrise, 2 day, 3 sunfall)
+// with fractional part as "position" in that part
+double get_day_time(double t);
 
 // handle modulo calculation for negative values the way I need it
 inline float myfmod(float a, float b) { return a-floor(a/b)*b; }
